Factor out some functionality to a HasType trait
